cli_automation: a sketch of how CLI automation will work
Related to a TMSU bug report, about the inefficiency of batch tagging/untagging
-- calling tmsu many times, instead of calling it once with multiple commands.

This sketch describes a scheme which reads NULL-separated commands,
with individual arguments separated by newlines.
It dispatches based on registration of subcommands -> the first argument in a set must be a subcommand.

To allow global options, the format includes a header.

WARNING: This has not been tested yet, use at your own risk.