core.connect() : new function, connects to and, if needed , initializes a TMSU db.
The connection object returned is a sqlite3.Connection subclass that includes a .normalize_path() method,
which should be used to normalize any paths you store or look up in the db.

core.relpath(database) : returns None if we are not using a .tmsu/db,
otherwise the path that files references by this DB should be relativized to.